feat: added solution for hash and modules
